refactor: 로컬스토리지가 클라이언트 사이드에서 동작해 발생하는 LCP 문제 개선(#72)
- 로컬 스토리지에 저장하던 배너 숨김 기간을 쿠키로 마이그레이션
